+++ Makefile
@@ -12,7 +12,14 @@ updates:
 
 .PHONY: test
 test:
-	@go test ./table
+	@go test -race -cover ./table
+
+.PHONY: test-coverage
+test-coverage: coverage.out
+	@go tool cover -html=coverage.out
+
+coverage.out: table/*.go go.*
+	@go test -coverprofile=coverage.out ./table
 
 .PHONY: fmt
 fmt:
